backend/requirements.txt
Flask
requests
flask-cors
python-dotenv  # If using .env
openai